[cmake] Improve build script
- Rely on SpaceVecAlg and sch-core targets
- Allow to build both Python 2/Python 3 at the same
- Add a test
